Add shooting to player controlled entity

+++ include/systems/RenderSystem.h
@@ -36,8 +36,8 @@ public:
             const auto texture = sprite->sprite.texture.get();
 
             SDL_Rect dstRect = {
-                static_cast<int>(transform->position.x - (sprite->fixed ? 0 : camera.x)),
-                static_cast<int>(transform->position.y - (sprite->fixed ? 0 : camera.y)),
+                static_cast<int>(transform->position.x - static_cast<float>(sprite->fixed ? 0 : camera.x)),
+                static_cast<int>(transform->position.y - static_cast<float>(sprite->fixed ? 0 : camera.y)),
                 static_cast<int>(static_cast<float>(sprite->rect.w) * transform->scale.x),
                 static_cast<int>(static_cast<float>(sprite->rect.h) * transform->scale.y),
             };
